WIP - many updates switch to IpAddr and remove auxip & replace wiht Endpoint, remove copy of same from transport
