+++ src/transport.rs
@@ -30,182 +30,14 @@ use std::str::FromStr;
 use rustc_serialize::{Decodable, Decoder, Encodable, Encoder};
 use utp::UtpListener;
 use std::fmt;
-use auxip;
 use connection::Connection;
 use std::io::BufReader;
 use util::ip_from_socketaddr;
 use util;
+use endpoint::{Endpoint, Port, Protocol};
 
 pub type Bytes = Vec<u8>;
 
-/// Enum representing supported transport protocols
-#[derive(Debug, PartialOrd, Ord, PartialEq, Eq, Hash, Clone, Copy)]
-pub enum Protocol {
-    /// TCP protocol
-    Tcp,
-    /// UTP protocol
-    Utp,
-}
-
-/// Enum representing endpoint of supported protocols
-#[derive(Debug, PartialEq, Eq, Hash, Clone, Copy)]
-pub enum Endpoint {
-    /// TCP endpoint
-    Tcp(SocketAddr),
-    /// UTP endpoint
-    Utp(SocketAddr),
-}
-
-impl Endpoint {
-    /// Construct a new Endpoint
-    pub fn new(addr: IpAddr, port: Port) -> Endpoint {
-        let socketaddr = match addr {
-            IpAddr::V4(a) => SocketAddr::V4(SocketAddrV4::new(a, port.number())),
-            IpAddr::V6(a) => SocketAddr::V6(SocketAddrV6::new(a, port.number(), 0, 0xe)),
-        };
-        match port {
-            Port::Tcp(_) => Endpoint::Tcp(socketaddr),
-            Port::Utp(_) => Endpoint::Utp(socketaddr),
-        }
-    }
-
-    /// Creates a Tcp(SocketAddr)
-    pub fn tcp<A: ToSocketAddrs>(addr: A) -> Endpoint {
-        match addr.to_socket_addrs().unwrap().next() {
-            Some(a) => Endpoint::Tcp(a),
-            None => panic!("Failed to parse valid IP address"),
-        }
-    }
-    /// Creates a Utp(SocketAddr)
-    pub fn utp<A: ToSocketAddrs>(addr: A) -> Endpoint {
-        match addr.to_socket_addrs().unwrap().next() {
-            Some(a) => Endpoint::Utp(a),
-            None => panic!("Failed to parse valid IP address"),
-        }
-    }
-    /// Returns SocketAddr.
-    pub fn get_address(&self) -> SocketAddr {
-        match *self {
-            Endpoint::Tcp(address) => address,
-            Endpoint::Utp(address) => address,
-        }
-    }
-
-    /// Returns port
-    pub fn get_port(&self) -> Port {
-        match *self {
-            Endpoint::Tcp(addr) => Port::Tcp(addr.port()),
-            Endpoint::Utp(addr) => Port::Utp(addr.port()),
-        }
-    }
-
-    /// Convert address format from Port to Endpoint
-    pub fn to_ip(&self) -> auxip::Endpoint {
-        let port = match self.get_port() {
-            Port::Tcp(n) => auxip::Port::Tcp(n),
-            Port::Utp(n) => auxip::Port::Udp(n),
-        };
-        auxip::Endpoint::new(ip_from_socketaddr(self.get_address()), port)
-    }
-
-    /// Check whether the current address is specified
-    /// returns true if address is un-specified, and false when specified
-    pub fn has_unspecified_ip(&self) -> bool {
-        util::is_unspecified(&ip_from_socketaddr(self.get_address()))
-    }
-
-    /// Convert address's format from ::std::net::IpAddr to Endpoint
-    pub fn map_ip_addr<F: Fn(IpAddr) -> IpAddr>(&self, f: F) -> Endpoint {
-        Endpoint::new(f(self.to_ip().ip()), self.get_port())
-    }
-}
-
-#[derive(Debug, RustcDecodable, RustcEncodable)]
-struct EndpointSerialiser {
-    pub protocol: String,
-    pub address: String,
-}
-
-impl Encodable for Endpoint {
-    fn encode<E: Encoder>(&self, e: &mut E) -> Result<(), E::Error> {
-        let s = EndpointSerialiser {
-            protocol: match *self {
-                Endpoint::Tcp(_) => "tcp".to_owned(),
-                Endpoint::Utp(_) => "utp".to_owned(),
-            },
-            address: self.get_address().to_string(),
-        };
-        try!(s.encode(e));
-        Ok(())
-    }
-}
-
-impl Decodable for Endpoint {
-    fn decode<D: Decoder>(d: &mut D) -> Result<Endpoint, D::Error> {
-        let decoded: EndpointSerialiser = try!(Decodable::decode(d));
-        match SocketAddr::from_str(&decoded.address) {
-            Ok(address) => {
-                if decoded.protocol == "tcp" {
-                    Ok(Endpoint::Tcp(address))
-                } else if decoded.protocol == "utp" {
-                    Ok(Endpoint::Utp(address))
-                } else {
-                    Err(d.error(&(format!("Unknown Protocol {}", decoded.protocol))))
-                }
-            }
-            _ => {
-                Err(d.error(&(format!("Expecting Protocol and SocketAddr string, but found : \
-                                       {:?}",
-                                      decoded))))
-            }
-        }
-    }
-}
-
-impl PartialOrd for Endpoint {
-    fn partial_cmp(&self, other: &Endpoint) -> Option<Ordering> {
-        Some(self.cmp(&other))
-    }
-}
-
-impl Ord for Endpoint {
-    fn cmp(&self, other: &Endpoint) -> Ordering {
-        use Endpoint::{Tcp, Utp};
-        match *self {
-            Tcp(ref a1) => {
-                match *other {
-                    Tcp(ref a2) => util::compare_ip_addrs(a1, a2),
-                    Utp(_) => Ordering::Greater,
-                }
-            }
-            Utp(ref a1) => {
-                match *other {
-                    Tcp(_) => Ordering::Less,
-                    Utp(ref a2) => util::compare_ip_addrs(a1, a2),
-                }
-            }
-        }
-    }
-}
-
-/// Enum representing port of supported protocols
-#[derive(Debug, PartialEq, Eq, Hash, Clone, RustcDecodable, RustcEncodable, Copy)]
-pub enum Port {
-    /// TCP port
-    Tcp(u16),
-    /// UTP port
-    Utp(u16),
-}
-
-impl Port {
-    /// Return the port
-    pub fn number(&self) -> u16 {
-        match *self {
-            Port::Tcp(p) => p,
-            Port::Utp(p) => p,
-        }
-    }
-}
 
 // --------------------------------------------------------------------
 
@@ -278,10 +110,12 @@ pub enum Receiver {
     Utp(cbor::Decoder<BufReader<utp_connections::UtpWrapper>>),
 }
 
+// TODO(dirvine) Figure why this unused allow is required  :09/01/2016
+#[allow(unused)]
 #[derive(Debug)]
 enum ReceiverError {
     FailedToDecodeCbor(cbor::CborError),
-    EndOfStream
+    EndOfStream,
 }
 
 impl fmt::Display for ReceiverError {
