+++ src/util.rs
@@ -20,7 +20,7 @@ use ip::IpAddr;
 use std::cmp::Ordering;
 use getifaddrs::{getifaddrs, IfAddr};
 use rustc_serialize::{Encodable, Decodable, Decoder, Encoder};
-use transport;
+use endpoint::{Endpoint, Port};
 use std::str::FromStr;
 
 #[cfg(test)]
@@ -152,9 +152,9 @@ pub fn ip_from_socketaddr(addr: SocketAddr) -> IpAddr {
     }
 }
 
-pub fn loopback_v4(port: transport::Port) -> transport::Endpoint {
+pub fn loopback_v4(port: Port) -> Endpoint {
     let ip = IpAddr::V4(Ipv4Addr::new(127, 0, 0, 1));
-    transport::Endpoint::new(ip, port)
+    Endpoint::new(ip, port)
 }
 
 pub fn is_v4(ip_addr: &IpAddr) -> bool {
@@ -319,12 +319,21 @@ pub fn heuristic_geo_cmp(ip1: &IpAddr, ip2: &IpAddr) -> Ordering {
 /// This function should really take IpAddr as an argument
 /// but it is used outside of this library and IpAddr
 /// is currently considered experimental.
-pub fn ifaddrs_if_unspecified(ep: transport::Endpoint) -> Vec<transport::Endpoint> {
-    if !is_unspecified(&ip_from_socketaddr(ep.get_address())) {
-        return vec![ep];
+pub fn ifaddrs_if_unspecified(ep: Endpoint) -> Vec<Endpoint> {
+    match ep.get_address() {
+        IpAddr::V4(ref addr) => {
+            if !::ip_info::v4::is_unspecified(addr) {
+                return vec![ep];
+            }
+        }
+        IpAddr::V6(ref addr) => {
+            if !::ip_info::v6::is_unspecified(addr) {
+                return vec![ep];
+            }
+        }
     }
 
-    let ep_is_v4 = is_v4(&ip_from_socketaddr(ep.get_address()));
+    let ep_is_v4 = is_v4(&ep.get_address());
 
     getifaddrs()
         .into_iter()
@@ -332,7 +341,7 @@ pub fn ifaddrs_if_unspecified(ep: transport::Endpoint) -> Vec<transport::Endpoin
             if ep_is_v4 != is_v4(&iface.addr) {
                 return None;
             }
-            Some(transport::Endpoint::new(iface.addr, ep.get_port()))
+            Some(Endpoint::new(iface.addr, ep.get_port()))
         })
         .collect()
 }
@@ -358,7 +367,7 @@ pub fn loopback_if_unspecified(addr: IpAddr) -> IpAddr {
 }
 
 #[cfg(test)]
-pub fn random_endpoint() -> ::transport::Endpoint {
+pub fn random_endpoint() -> Endpoint {
     // TODO - randomise V4/V6 and TCP/UTP
     let address =
         ::std::net::SocketAddrV4::new(::std::net::Ipv4Addr::new(::rand::random::<u8>(),
@@ -366,11 +375,11 @@ pub fn random_endpoint() -> ::transport::Endpoint {
                                                                 ::rand::random::<u8>(),
                                                                 ::rand::random::<u8>()),
                                       ::rand::random::<u16>());
-    ::transport::Endpoint::Tcp(::std::net::SocketAddr::V4(address))
+    Endpoint::Tcp(::std::net::SocketAddr::V4(address))
 }
 
 #[cfg(test)]
-pub fn random_endpoints(count: usize) -> Vec<::transport::Endpoint> {
+pub fn random_endpoints(count: usize) -> Vec<Endpoint> {
     let mut contacts = Vec::new();
     for _ in 0..count {
         contacts.push(random_endpoint());
@@ -379,7 +388,7 @@ pub fn random_endpoints(count: usize) -> Vec<::transport::Endpoint> {
 }
 
 #[cfg(test)]
-pub fn random_global_endpoint() -> ::transport::Endpoint {
+pub fn random_global_endpoint() -> Endpoint {
     // TODO - randomise V4/V6 and TCP/UTP
     let address =
         ::std::net::SocketAddrV4::new(::std::net::Ipv4Addr::new(173, // ensure is a global addr
@@ -387,11 +396,11 @@ pub fn random_global_endpoint() -> ::transport::Endpoint {
                                                                 ::rand::random::<u8>(),
                                                                 ::rand::random::<u8>()),
                                       ::rand::random::<u16>());
-    ::transport::Endpoint::Tcp(::std::net::SocketAddr::V4(address))
+    Endpoint::Tcp(::std::net::SocketAddr::V4(address))
 }
 
 #[cfg(test)]
-pub fn random_global_endpoints(count: usize) -> Vec<::transport::Endpoint> {
+pub fn random_global_endpoints(count: usize) -> Vec<Endpoint> {
     let mut contacts = Vec::new();
     for _ in 0..count {
         contacts.push(random_global_endpoint());
